+++ annes_inquiry/app/helpers/annes_inquiry/form_helper.rb
@@ -21,7 +21,7 @@ module AnnesInquiry
     end
 
     def inquiry_field_attributes(field, input, scope:)
-      id = "#{scope}_#{field.key}"
+      id = "#{scope}_#{field.key}".gsub(/[^a-zA-Z0-9_-]/, "_")
       { id: id, required: field.required, placeholder: field.placeholder,
         aria: { describedby: "#{id}_help #{id}_errors", invalid: input.errors[field.key].any? ? "true" : nil } }
     end